+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Change Log
 
 .. There should always be an "Unreleased" section for changes pending release.
 
+Unreleased
+**********
+* Add content cache
+
 1.5.0 - 2023-10-18
 ******************
 * Add management command to generate course prompts